requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
bard
beautifulsoup4
bing_image_downloader
bs4
telegraph
wget
unidecode
cloudscraper
curl-cffi
deepai
dnspython
emojis
Faker
ffmpeg-python
future
g4f
geopy
gitpython
google-api-python-client
gpytranslate
gtts
hachoir
heroku3
httpx[http2]==0.27.2
img2pdf
instaloader
langdetect
lexica-api
motor
nekosbest
numpy
openai==0.28.0
opencv-python
pickledb
pillow==10.3.0
psutil
ntgcalls
py-tgcalls==0.9.7  # Only one version of py-tgcalls is specified
pycountry
pycryptodome
pydub
pyfiglet
pykeyboard
pyshorteners
python-dotenv
python-telegram-bot
python-whois
pyttsx3
pytube
pytz
pyyaml
qrcode
requests
search_engine_parser
speedtest-cli
SpeechRecognition
spotipy
stripe==10.7.0
mysql-connector-python==9.0.0
telebot==0.0.5
youtube_dl==2021.12.17
git+https://github.com/DAXXTEAM/daxxhub
python-telegram-bot==21.1.1
pyyaml==6.0.1
MukeshAPI
youtube-search
youtube-search-python
git+https://github.com/KurimuzonAkuma/pyrogram.git
git+https://github.com/yt-dlp/yt-dlp.git
git+https://github.com/AsmSafone/SafoneAPI
